EGMC
Added all parking and changed the taxiway layout as to the present day arrangement. The airport was changed about four or five years ago because flights with easyjet and flybe were going to start. The current layout in IF shows what the airport was like when only small aircraft used to fly to and from Southend.  
I have added some car parks and pedestrianised areas, but may add some more around the airport at a later date.