EHBX Heliport Fix
Fixed this heliport. Used to have a very short runway meant to be a
heliport, deleted runway and windsocks, added heloport. (This apt.dat
file has literally nothing else in it except that helipad object,
probably the smallest apt.dat file in the entire repo. Why are we even
putting this here)