KLGA – Few Additions
Added a few transient parking spots as seen on the satellite imagery.
Quickly fixed an odd gate and snapped a couple of things.

Airport is in major need of an overhaul. Lots of things are unsnapped,
etc. I will try to get to it at some point in the future if no one
beats me to the punch. :)